DLS-1417 - Changed sba claim to an int

+++ src/main/scala/uk/gov/hmrc/ct/computations/SBA01.scala
@@ -41,5 +41,5 @@ case class Building(
                      earliestWrittenContract: Option[LocalDate],
                      nonResidentialActivityStart: Option[LocalDate],
                      cost: Option[BigDecimal],
-                     claim: Option[BigDecimal]
+                     claim: Option[Int]
                    )
